pyproject.toml
[tool.poetry]
name = "hard-train"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
drf-yasg = "^1.21.8"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.9.10"
whitenoise = "^6.8.2"
flower = "^2.0.1"
redis = "^5.2.0"
pika = "^1.3.2"
django-filter = "^24.3"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
pytest = "^8.3.4"
pytest-django = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
